+++ db/pvault.sql
@@ -93,7 +93,8 @@ CREATE TABLE `formulasMetaData` (
   `isMade` INT NOT NULL DEFAULT '0',
   `madeOn` DATETIME NULL DEFAULT NULL,
   `customer_id` INT NOT NULL DEFAULT '0',
-  `status` INT NOT NULL DEFAULT '0'
+  `status` INT NOT NULL DEFAULT '0',
+  `toDo` INT NOT NULL DEFAULT '0'
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_general_ci;
 
 CREATE TABLE `IFRALibrary` (