+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # CHANGELOG
 ### Version 5.7
 - Added the option to attach documents in formulas
+- Formula making view improvements
 
 ### Version 5.6
 - HTML Templates added