+++ releasenotes.md
@@ -1,3 +1,4 @@
 Whats New:
 ----------------------
 - Added the option to attach documents in formulas
+- Formula making view improvements